+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Configuration file
+- Logging to file
 ### Changed
 - Renamed SshBasedConnection to Connection
 - Connection now accepts any instance of Transport
+- -c key is now shorthand for --config, not --chat-id
 
 ## 0.1.0 - 2020-03-26
 ### Added